refactor: split analysis.ts into multiple files
